PDFBOX-5889: support long values for COSInteger
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1921496 13f79535-47bb-0310-9956-ffa450edef68

+++ debugger/src/main/java/org/apache/pdfbox/debugger/ui/PDFTreeCellRenderer.java
@@ -134,7 +134,7 @@ public class PDFTreeCellRenderer extends DefaultTreeCellRenderer
         }
         else if (nodeValue instanceof COSInteger)
         {
-            result = Integer.toString(((COSInteger) nodeValue).intValue());
+            result = Long.toString(((COSInteger) nodeValue).longValue());
         }
         else if (nodeValue instanceof COSString)
         {
